+++ src/dflow/context.py
@@ -1,0 +1,7 @@
+class Context(object):
+    """
+    Context
+    :method get_annotations: 
+    """
+    def get_annotations(self):
+        return {}