+++ src/dflow/resource.py
@@ -1,5 +1,3 @@
-import yaml
-
 class Resource(object):
     """
     Resource
@@ -13,32 +11,3 @@ class Resource(object):
     failure_condition = None
     def get_manifest(self, command, script):
         raise NotImplementedError()
-
-class SlurmJob(Resource):
-    def __init__(self, header="", node_selector=None, prepare=None, results=None):
-        self.header = header
-        self.action = "create"
-        self.success_condition = "status.status == Succeeded"
-        self.failure_condition = "status.status == Failed"
-        self.node_selector = node_selector
-        self.prepare = prepare
-        self.results = results
-
-    def get_manifest(self, command, script, workdir="."):
-        manifest = {
-            "apiVersion": "wlm.sylabs.io/v1alpha1",
-            "kind": "SlurmJob",
-            "metadata": {
-                "name": "{{pod.name}}"
-            },
-            "spec": {
-                "batch": self.header + "\ncd %s\ncat <<EOF | %s\n%s\nEOF" % (workdir, " ".join(command), script)
-            }
-        }
-        if self.node_selector is not None:
-            manifest["spec"]["nodeSelector"] = self.node_selector
-        if self.prepare is not None:
-            manifest["spec"]["prepare"] = self.prepare
-        if self.results is not None:
-            manifest["spec"]["results"] = self.results
-        return yaml.dump(manifest)